uio: uio_pdrv_genirq: Remove MODULE_DEVICE_TABLE
The struct of_device_id list here is filled in at load time. It does not
work with MODULE_DEVICE_TABLE, which generates an alias at build time.
In this case, it generates the aliases "of:N*T*" and "of:N*T*C*", which
matches *any* OF device node, which is obviously not intended. It
confuses userspace into loading this module for any OF device, and
confuses anyone trying to resolve an OF modalias.

Fixes: 27760f868663 ("uio: uio_pdrv_genirq: Add OF support")
Signed-off-by: Vivian Wang <wangruikang@iscas.ac.cn>